Adding monitoring priority to all triggers.
Triggers now have an associated monitoring priority. This priority is used
to determine whether or not to send StatsD metrics for a trigger. It is
persisted in the database and can be accessed via a model. Accordingly,
it can be used for additional monitoring beyond the monitoring built into
Piezo. For example, a monitoring job could be created and the priority used
to determine whether an alert should be created, and what priority it
should be when a trigger enters into an error state or fails to run.