Add a maximum amount of time between successes for each trigger.
This is a continuation of the monitoring work done by adding monitoring
priority. This adds an SLA for each trigger - the maximum number of time
that can pass between successes. We have a Job at Lucidchart which enforces
this internally. Including it in the Piezo project requires us to open
source an additional project. Until then, you can write a job that consumes
this information and alerts when a trigger is in an error state.