Guarantees that scheduler gets shut down on SIGTERM
We use upstart to manager our quartz jobs. Upstart sends SIGTERM to
stop the upstart job. Without this change, this terminates the jobs
without calling scheduler.shutdown(true). I think this will fix the
issues we have of triggers being set to run in the past. If not, it's
still a more clean shutdown of quartz.